FP-10: Moved utils/types to different folders
