+++ assets/components/form/SliderWidget.tsx
@@ -1,6 +1,6 @@
 import { Form } from "@/types/forms/Form"
 import { FormWidgetProps } from "@/types/props/FormWidgetProps"
-import { validateFormWidgetProps } from "@/util/validateFormWidgetProps"
+import { validateFormWidgetProps } from "@/util/forms/validateFormWidgetProps"
 import { Slider } from "@mui/material"
 import React, { ReactElement } from "react"
