--- assets/types/WidgetOption.ts
@@ -1,4 +1,0 @@
-import { RadioOption } from "@/types/RadioOption"
-import { SelectOption } from "@/types/SelectOption"
-
-export type WidgetOption = SelectOption | RadioOption
