+++ assets/types/options/UserGroupOption.ts
@@ -1,6 +1,6 @@
-import { ModelOption } from '@/types/ModelOption'
+import { ModelOption } from '@/types/options/ModelOption'
 import { UserGroupModel } from '@/types/UserGroupModel'
-import { RadioOption } from '@/types/RadioOption'
+import { RadioOption } from '@/types/options/RadioOption'
 
 export class UserGroupOption extends ModelOption<UserGroupModel, RadioOption> {
     getOption(): RadioOption {
