+++ assets/types/options/UserOption.ts
@@ -1,5 +1,5 @@
-import { ModelOption } from '@/types/ModelOption'
-import { SelectOption } from '@/types/SelectOption'
+import { ModelOption } from '@/types/options/ModelOption'
+import { SelectOption } from '@/types/options/SelectOption'
 import { UserModel } from '@/types/UserModel'
 
 export class UserOption extends ModelOption<UserModel, SelectOption> {
