+++ assets/types/options/WidgetOption.ts
@@ -1,0 +1,4 @@
+import { RadioOption } from "@/types/options/RadioOption"
+import { SelectOption } from "@/types/options/SelectOption"
+
+export type WidgetOption = SelectOption | RadioOption
